build: added devel build target
Added a development build target.  Like the staging build, this
installs an ssh server, but unlike the staging build, this pulls
packages from the production package repo.

Also, install git, pip3, npm and zile packages on the development
and staging builds.  This reduces the setup necessary to do
development on epipylon from an installed image.

Also, we need to specify the size limit when mounting the boot
partition to avoid operlapping mount regions with the root partition.
Otherwise, Fedora 26 will disallow the mount operation.  Fedora 25
allowed it.